refactor: profile 닉네임 관련 기능 member 패키지로 이동
