+++ backend/sokdak/src/main/java/com/wooteco/sokdak/member/controller/MemberController.java
@@ -1,15 +1,20 @@
 package com.wooteco.sokdak.member.controller;
 
+import com.wooteco.sokdak.auth.dto.AuthInfo;
 import com.wooteco.sokdak.member.dto.EmailRequest;
 import com.wooteco.sokdak.member.dto.SignupRequest;
 import com.wooteco.sokdak.member.dto.UniqueResponse;
 import com.wooteco.sokdak.member.dto.VerificationRequest;
 import com.wooteco.sokdak.member.service.EmailService;
 import com.wooteco.sokdak.member.service.MemberService;
+import com.wooteco.sokdak.profile.dto.NicknameResponse;
+import com.wooteco.sokdak.profile.dto.NicknameUpdateRequest;
+import com.wooteco.sokdak.support.token.Login;
 import javax.validation.Valid;
 import org.springframework.http.HttpStatus;
 import org.springframework.http.ResponseEntity;
 import org.springframework.web.bind.annotation.GetMapping;
+import org.springframework.web.bind.annotation.PatchMapping;
 import org.springframework.web.bind.annotation.PostMapping;
 import org.springframework.web.bind.annotation.RequestBody;
 import org.springframework.web.bind.annotation.RequestMapping;
@@ -56,5 +61,18 @@ public class MemberController {
     public ResponseEntity<Void> signUp(@Valid @RequestBody SignupRequest signupRequest) {
         memberService.signUp(signupRequest);
         return ResponseEntity.status(HttpStatus.CREATED).build();
+    }
+
+    @GetMapping("/nickname")
+    public ResponseEntity<NicknameResponse> findNickname(@Login AuthInfo authInfo) {
+        NicknameResponse nicknameResponse = memberService.findNickname(authInfo);
+        return ResponseEntity.ok(nicknameResponse);
+    }
+
+    @PatchMapping("/nickname")
+    public ResponseEntity<Void> editNickname(@RequestBody NicknameUpdateRequest nicknameUpdateRequest,
+                                             @Login AuthInfo authInfo) {
+        memberService.editNickname(nicknameUpdateRequest, authInfo);
+        return ResponseEntity.noContent().build();
     }
 }
