+++ backend/sokdak/src/main/java/com/wooteco/sokdak/member/service/MemberService.java
@@ -1,14 +1,20 @@
 package com.wooteco.sokdak.member.service;
 
+import com.wooteco.sokdak.auth.dto.AuthInfo;
 import com.wooteco.sokdak.auth.service.AuthService;
 import com.wooteco.sokdak.auth.service.Encryptor;
 import com.wooteco.sokdak.member.domain.Member;
+import com.wooteco.sokdak.member.domain.Nickname;
 import com.wooteco.sokdak.member.dto.SignupRequest;
 import com.wooteco.sokdak.member.dto.UniqueResponse;
 import com.wooteco.sokdak.member.dto.VerificationRequest;
 import com.wooteco.sokdak.member.exception.InvalidSignupFlowException;
+import com.wooteco.sokdak.member.exception.MemberNotFoundException;
 import com.wooteco.sokdak.member.exception.PasswordConfirmationException;
 import com.wooteco.sokdak.member.repository.MemberRepository;
+import com.wooteco.sokdak.profile.dto.NicknameResponse;
+import com.wooteco.sokdak.profile.dto.NicknameUpdateRequest;
+import com.wooteco.sokdak.profile.exception.DuplicateNicknameException;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
 
@@ -90,5 +96,27 @@ public class MemberService {
         if (!password.equals(passwordConfirmation)) {
             throw new PasswordConfirmationException();
         }
+    }
+
+    @Transactional
+    public void editNickname(NicknameUpdateRequest nicknameUpdateRequest, AuthInfo authInfo) {
+        Member member = memberRepository.findById(authInfo.getId())
+                .orElseThrow(MemberNotFoundException::new);
+
+        Nickname validNickname = new Nickname(nicknameUpdateRequest.getNickname());
+        validateUniqueNickname(validNickname);
+        member.updateNickname(validNickname);
+    }
+
+    private void validateUniqueNickname(Nickname validNickname) {
+        if (memberRepository.existsMemberByNickname(validNickname)) {
+            throw new DuplicateNicknameException();
+        }
+    }
+
+    public NicknameResponse findNickname(AuthInfo authInfo) {
+        Member member = memberRepository.findById(authInfo.getId())
+                .orElseThrow(MemberNotFoundException::new);
+        return NicknameResponse.of(member);
     }
 }
