+++ backend/sokdak/src/test/java/com/wooteco/sokdak/member/controller/MemberControllerTest.java
@@ -1,9 +1,14 @@
 package com.wooteco.sokdak.member.controller;
 
+import static com.wooteco.sokdak.util.fixture.MemberFixture.AUTH_INFO;
 import static org.hamcrest.Matchers.equalTo;
 import static org.mockito.ArgumentMatchers.any;
+import static org.mockito.ArgumentMatchers.refEq;
 import static org.mockito.BDDMockito.given;
+import static org.mockito.Mockito.doNothing;
+import static org.mockito.Mockito.doReturn;
 import static org.mockito.Mockito.doThrow;
+import static org.springframework.http.HttpHeaders.AUTHORIZATION;
 import static org.springframework.restdocs.mockmvc.MockMvcRestDocumentation.document;
 
 import com.wooteco.sokdak.member.dto.EmailRequest;
@@ -11,10 +16,14 @@ import com.wooteco.sokdak.member.dto.SignupRequest;
 import com.wooteco.sokdak.member.dto.UniqueResponse;
 import com.wooteco.sokdak.member.dto.VerificationRequest;
 import com.wooteco.sokdak.member.exception.InvalidAuthCodeException;
+import com.wooteco.sokdak.member.exception.InvalidNicknameException;
 import com.wooteco.sokdak.member.exception.InvalidSignupFlowException;
 import com.wooteco.sokdak.member.exception.NotWootecoMemberException;
 import com.wooteco.sokdak.member.exception.PasswordConfirmationException;
 import com.wooteco.sokdak.member.exception.TicketUsedException;
+import com.wooteco.sokdak.profile.dto.NicknameResponse;
+import com.wooteco.sokdak.profile.dto.NicknameUpdateRequest;
+import com.wooteco.sokdak.profile.exception.DuplicateNicknameException;
 import com.wooteco.sokdak.util.ControllerTest;
 import javax.transaction.Transactional;
 import org.junit.jupiter.api.DisplayName;
@@ -304,5 +313,92 @@ class MemberControllerTest extends ControllerTest {
 
                 .apply(document("member/signup/fail/duplicateUsername"))
                 .statusCode(HttpStatus.BAD_REQUEST.value());
+    }
+
+    @DisplayName("닉네임 조회 시 200 반환")
+    @Test
+    void findNickname() {
+        setUpArgumentResolver();
+        doReturn(new NicknameResponse("chrisNickname"))
+                .when(memberService)
+                .findNickname(any());
+
+        restDocs
+                .contentType(MediaType.APPLICATION_JSON_VALUE)
+                .header(AUTHORIZATION, "Bearer chris")
+                .when().get("/members/nickname")
+                .then().log().all()
+                .assertThat()
+                .apply(document("member/find/nickname/success"))
+                .statusCode(HttpStatus.OK.value());
+    }
+
+    @DisplayName("닉네임 변경 시 204 반환")
+    @Test
+    void editNickname() {
+        setUpArgumentResolver();
+        NicknameUpdateRequest nicknameUpdateRequest = new NicknameUpdateRequest("chrisNick2");
+        doNothing()
+                .when(memberService)
+                .editNickname(any(), any());
+
+        restDocs
+                .contentType(MediaType.APPLICATION_JSON_VALUE)
+                .header(AUTHORIZATION, "Bearer chris")
+                .body(nicknameUpdateRequest)
+                .when().patch("/members/nickname")
+                .then().log().all()
+                .assertThat()
+                .apply(document("member/patch/nickname/success"))
+                .statusCode(HttpStatus.NO_CONTENT.value());
+    }
+
+    @DisplayName("닉네임 변경 시 이미 있는 닉네임이면 400 반환")
+    @Test
+    void editNickname_Exception_Duplicate() {
+        setUpArgumentResolver();
+        NicknameUpdateRequest nicknameUpdateRequest = new NicknameUpdateRequest("hunchNickname");
+        doThrow(new DuplicateNicknameException())
+                .when(memberService)
+                .editNickname(refEq(nicknameUpdateRequest), any());
+
+        restDocs
+                .contentType(MediaType.APPLICATION_JSON_VALUE)
+                .header(AUTHORIZATION, "Bearer chris")
+                .body(nicknameUpdateRequest)
+                .when().patch("/members/nickname")
+                .then().log().all()
+                .assertThat()
+                .apply(document("member/patch/nickname/fail/duplicate"))
+                .statusCode(HttpStatus.BAD_REQUEST.value());
+    }
+
+    @DisplayName("닉네임 변경 시 잘못된 형식이면 400 반환")
+    @Test
+    void editNickname_Exception_InvalidFormat() {
+        setUpArgumentResolver();
+        NicknameUpdateRequest nicknameUpdateRequest = new NicknameUpdateRequest("");
+        doThrow(new InvalidNicknameException())
+                .when(memberService)
+                .editNickname(refEq(nicknameUpdateRequest), any());
+
+        restDocs
+                .contentType(MediaType.APPLICATION_JSON_VALUE)
+                .header(AUTHORIZATION, "Bearer chris")
+                .body(nicknameUpdateRequest)
+                .when().patch("/members/nickname")
+                .then().log().all()
+                .assertThat()
+                .apply(document("member/patch/nickname/fail/invalidFormat"))
+                .statusCode(HttpStatus.BAD_REQUEST.value());
+    }
+
+    private void setUpArgumentResolver() {
+        doReturn(true)
+                .when(authInterceptor)
+                .preHandle(any(), any(), any());
+        doReturn(AUTH_INFO)
+                .when(authenticationPrincipalArgumentResolver)
+                .resolveArgument(any(), any(), any(), any());
     }
 }
